+++ api/src/main/java/com/grash/model/PartConsumption.java
@@ -17,7 +17,7 @@ import javax.validation.constraints.NotNull;
 public class PartConsumption extends CompanyAudit {
     @NotNull
     @Min(value = 0L, message = "The value must be positive")
-    private int quantity;
+    private double quantity;
 
     @ManyToOne
     @NotNull
@@ -28,13 +28,13 @@ public class PartConsumption extends CompanyAudit {
     @OnDelete(action = OnDeleteAction.CASCADE)
     private WorkOrder workOrder;
 
-    public PartConsumption(Part part, WorkOrder workOrder, int quantity) {
+    public PartConsumption(Part part, WorkOrder workOrder, double quantity) {
         this.part = part;
         this.workOrder = workOrder;
         this.quantity = quantity;
     }
 
-    public long getCost() {
+    public double getCost() {
         return part.getCost() * quantity;
     }
 }
